Alright cleaned this up so that the YJster can take a look
